7th round

+++ src/main/java/model/Calculator.java
@@ -43,7 +43,7 @@ public class Calculator {
         if (result > Integer.MAX_VALUE || result < Integer.MIN_VALUE) {
             throw new CalculatorException("Превышен порог значений");
         }
-        return String.valueOf(result);
+        return String.valueOf(BigDecimal.valueOf(result));
     }
